+++ magda-web-client/src/Components/Home/HomePage.scss
@@ -10,7 +10,7 @@
         width: 100%;
         height: 100vh;
         z-index: -1;
-        min-width: 300px;
+        min-width: $minimum;
     }
 
     .homepage-background {
@@ -25,7 +25,7 @@
         right: 0;
         bottom: 0;
         z-index: -1;
-        min-width: 300px;
+        min-width: $minimum;
     }
 
     .homepage-tagline {
