+++ magda-web-client/src/Routes.tsx
@@ -62,8 +62,15 @@ const HighlightsAdminPage = makeAsync(() =>
 const HomeAdminPage = makeAsync(() =>
     import("Components/Home/HomeAdminPage").then(module => module.default)
 );
+
 const DatasetAddPage = makeAsync(() =>
     import("Components/Dataset/DatasetAddPage").then(module => module.default)
+);
+
+const DatasetAddFilesPage = makeAsync(() =>
+    import("Components/Dataset/DatasetAddFilesPage").then(
+        module => module.default
+    )
 );
 const StaticPage = makeAsync(() =>
     import("Components/Static/StaticPage").then(module => module.default)
@@ -206,7 +213,56 @@ const Routes = () => {
             {config.featureFlags.cataloguing && (
                 <Route
                     exact
-                    path="/dataset/new"
+                    path="/dataset/add"
+                    component={withHeader(DatasetAddPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/files"
+                    component={withHeader(DatasetAddFilesPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/urls"
+                    component={withHeader(DatasetAddPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/metadata/1"
+                    component={withHeader(DatasetAddPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/metadata/2"
+                    component={withHeader(DatasetAddPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/metadata/3"
+                    component={withHeader(DatasetAddPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/metadata/4"
+                    component={withHeader(DatasetAddPage, false)}
+                />
+            )}
+            {config.featureFlags.cataloguing && (
+                <Route
+                    exact
+                    path="/dataset/add/bulk"
                     component={withHeader(DatasetAddPage, false)}
                 />
             )}
