+++ magda-web-client/src/_variables.scss
@@ -47,6 +47,7 @@ $magda-yellow: #f1c40f;
 
 $banner-text: $AU-color-foreground-text;
 
+$minimum: 576px;
 $small: 768px;
 $medium: 992px;
 $large: 1200px;
