+++ magda-web-client/src/index.scss
@@ -4,7 +4,7 @@ html {
 }
 body {
     overflow-y: scroll;
-    min-width: 300px;
+    min-width: $minimum;
 }
 .au-body .au-grid {
     p,
